fix: type

+++ src/dishes/entities/dishes.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity } from 'typeorm';
 import { GeneralEntity } from '@utils/base.entity';
 import { DISHES_TYPE } from '../enums/dishes.enum';
 
-@Entity({name: 'dishes'})
+@Entity({ name: 'dishes' })
 export class DishesEntity extends GeneralEntity {
   @Column({ type: 'varchar', name: 'name' })
   name: string;
@@ -13,12 +13,12 @@ export class DishesEntity extends GeneralEntity {
   @Column({ type: 'enum', name: 'type', enum: DISHES_TYPE })
   type: DISHES_TYPE;
 
-  @Column({ type: 'smallint', name: 'time' })
+  @Column({ type: 'decimal', name: 'time' })
   time: number;
 
-  @Column({ type: 'decimal', name: 'mark', precision: 6, scale: 2  })
+  @Column({ type: 'decimal', name: 'mark', precision: 6, scale: 2 })
   mark: number;
 
-  @Column({ type: 'decimal', name: 'price', precision: 6, scale: 2  })
+  @Column({ type: 'decimal', name: 'price', precision: 6, scale: 2 })
   price: number;
 }
